Formatting changes, added Queen movement (testing pending) and cleaned up code

+++ include/pieces/King.hpp
@@ -10,7 +10,7 @@ class King
 
     static void initialize();
     static std::array<std::uint64_t, 64> generateMoves();
-    static void visDebug(int pos);
+    static void visDebug(unsigned long pos);
 
   private:
     static std::array<std::uint64_t, 64> moves;
