+++ include/pieces/Rook.hpp
@@ -10,7 +10,7 @@ public:
 
   static void initialize();
   static std::array<std::uint64_t, 64> generateMoves();
-  static void visDebug(int pos);
+  static void visDebug(unsigned long pos);
   const static std::array<std::uint64_t, 64> getMoves() { return moves; }
 
 private:
